+++ Editor/SceneCollection_Editor.cs
@@ -14,7 +14,6 @@
 // *   See the License for the specific language governing permissions and
 // *   limitations under the License.
 
-
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEditor;
@@ -27,16 +26,36 @@ namespace HH.MultiSceneToolsEditor
     public class SceneCollection_Editor : Editor
     {
         SceneCollection script;
+        SerializedProperty _Title;
+        SerializedProperty _ActiveSceneIndex;
+        SerializedProperty _Scenes;
+        
 
         private void OnEnable()
         {
             script = target as SceneCollection;
+
+            _Title = serializedObject.FindProperty("Title");
+            _ActiveSceneIndex = serializedObject.FindProperty("ActiveSceneIndex");
+            _Scenes = serializedObject.FindProperty("Scenes");
         }
 
         public override void OnInspectorGUI()
         {
-            base.OnInspectorGUI();
-        
+            serializedObject.Update();
+
+            EditorGUILayout.PropertyField(_Title);
+
+            if(_Scenes.isExpanded)
+                EditorGUILayout.PropertyField(_Scenes, new GUIContent("Active Scene | In Build | Target Scene"));
+            else
+                EditorGUILayout.PropertyField(_Scenes);
+
+            if(_Scenes.serializedObject.hasModifiedProperties)
+                updateActiveSceneIndex();
+
+            serializedObject.ApplyModifiedProperties();
+
             if(GUILayout.Button("Load Collection"))
             {
                 script.LoadCollection();
@@ -70,6 +89,28 @@ namespace HH.MultiSceneToolsEditor
                 return true;
             }
             return false;
+        }
+
+        void updateActiveSceneIndex()
+        {
+            bool isUpdated = false;
+            for (int i = 0; i < _Scenes.arraySize; i++)
+            {
+                if(!_Scenes.GetArrayElementAtIndex(i).FindPropertyRelative("IsActive").boolValue)
+                    continue;
+
+                if(i == _ActiveSceneIndex.intValue)
+                    continue;
+
+                if(_ActiveSceneIndex.intValue >= 0)
+                    _Scenes.GetArrayElementAtIndex(_ActiveSceneIndex.intValue).FindPropertyRelative("IsActive").boolValue = false;
+
+                _ActiveSceneIndex.intValue = i;
+                isUpdated = true;
+                break;
+            }
+            if(!isUpdated)
+                _ActiveSceneIndex.intValue = -1;
         }
     }
 }